fix(esp_i2s_source): identify stations by stable id, not array index
GET /api/stations exposes each station's stable id and the web UI sends
that id back for delete/move/edit, but stations_remove/move/update treated
the ?id= param as an array index. Once ids diverged from indices (after any
delete/reorder) the UI silently targeted the wrong station or failed.

Take a uint32_t stable id in those three functions and resolve it to an
index via the existing station_store_index_by_id() under the store mutex
(atomic; unknown id -> not found). Verified live: delete and move by stable
id now hit the correct station.

+++ esp_i2s_source/components/web_ui/web_ui_radio.c
@@ -236,7 +236,7 @@ esp_err_t stations_put_h(httpd_req_t *req)
         httpd_req_get_url_query_str(req, q, sizeof(q)) == ESP_OK &&
         httpd_query_key_value(q, "move", mv, sizeof(mv)) == ESP_OK) {
         int delta = !strcmp(mv, "up") ? -1 : (!strcmp(mv, "down") ? 1 : 0);
-        esp_err_t err = delta != 0 ? stations_move(id, delta) : ESP_ERR_INVALID_ARG;
+        esp_err_t err = delta != 0 ? stations_move((uint32_t)id, delta) : ESP_ERR_INVALID_ARG;
         if (err == ESP_ERR_INVALID_ARG) {
             station_reply_bad(req, "cannot reorder (already at edge or bad id)");
         } else if (err != ESP_OK) {
@@ -252,7 +252,7 @@ esp_err_t stations_put_h(httpd_req_t *req)
         return ESP_OK;
     }
     station_result_t reason = STATION_ERR_INVALID_ARG;
-    esp_err_t err = stations_update(id, name, url, &reason);
+    esp_err_t err = stations_update((uint32_t)id, name, url, &reason);
     if (err == ESP_ERR_INVALID_ARG) {
         station_reply_bad(req, station_result_str(reason));
     } else if (err != ESP_OK) {
@@ -269,7 +269,7 @@ esp_err_t stations_delete_h(httpd_req_t *req)
     if (guard != ESP_OK) return guard;
 
     int id = station_id_param(req);
-    esp_err_t err = id >= 0 ? stations_remove(id) : ESP_ERR_INVALID_ARG;
+    esp_err_t err = id >= 0 ? stations_remove((uint32_t)id) : ESP_ERR_INVALID_ARG;
     if (err == ESP_ERR_INVALID_ARG) {
         station_reply_bad(req, "station not found");
     } else if (err != ESP_OK) {
